+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: iryoga <iryoga@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/11 06:05:22 by iryoga            #+#    #+#             */
-/*   Updated: 2022/08/18 07:44:15 by iryoga           ###   ########.fr       */
+/*   Updated: 2022/08/18 07:52:16 by iryoga           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,6 @@ int		ft_print_str(char *s);
 int		ft_print_address(uintptr_t ptr);
 int		ft_print_nbr(int n);
 int		ft_print_unnbr(unsigned int n);
-int		ft_print_hex(unsigned long long n, int specifier);
+int		ft_print_hex(unsigned int n, int specifier);
 
 #endif
